Revert SkSequentialChatMessageContentConnector
